decided to take a different approach
because things weren't working right and it wasn't mapping out each character the way I had it. Looked over what I did yesterday and decided to simplify by making the character a const rather than a class and then in the App.js using a map arrow function because that seems to be the only way I can get these things to display (maybe that is the only correct way to do this). Anyway, everything displays now but it's pretty ugly and doesn't look like cards or anything. So back to work formatting everything.