Figured out a way to iterate over the different lists
Unfortunately, because it's only the first page of the different lists, the references don't line up the way that I hoped it would
I'll probably come back to it and try to figure out a way to fetch the data from the different lists specifically instead of what currently is happening.
Until then, I'm going to work on getting my cards to display again.