Nothing done
I was completely lost on this, mainly due to the online database thing, I didn't understand how we would get the info from it, and just mass confusion went on. Nothing is finshed, I attempted to try adn do it like one of the code sandbox things but this didn't work either.