MVP
"build a react app that displays this fun Star Wars data. Build a card for each object in the starwars array."

data that wa fetched from the API is being displayed in the app. So in this regard the sprint has reached MVP, but I will add more styling and attempt the stretch