Final styling
I just changed the layout of the cards on the page. Wasn't really sure what else to do with the data in the array. I attempted to display a single card at a time, but ran into an issue that doesn't make sense to me. Even though I was passing an object as a prop, I couldn't access the values by referencing the key with dot notation. It worked when inside App.js, but returned undefined when passed to the child component. React dev-tools had the props as an object as expected, but when referenced each value came back as undefined.